+++ client/include/autd3/core/osal_timer/win32/timer.hpp
@@ -3,7 +3,7 @@
 // Created Date: 01/05/2021
 // Author: Shun Suzuki
 // -----
-// Last Modified: 04/07/2021
+// Last Modified: 23/07/2021
 // Modified By: Shun Suzuki (suzuki@hapis.k.u-tokyo.ac.jp)
 // -----
 // Copyright (c) 2021 Hapis Lab. All rights reserved.
@@ -26,7 +26,7 @@ template <typename T>
 class Timer {
  public:
   Timer(std::unique_ptr<T> handler, const uint32_t timer_id) : _handler(std::move(handler)), _timer_id(timer_id), _is_closed(false) {}
-  ~Timer() { (void)this->stop(); }
+  ~Timer() { const auto _ = this->stop(); }
 
   [[nodiscard]] static std::unique_ptr<Timer> start(std::unique_ptr<T> handler, uint32_t interval_us) {
     interval_us = interval_us / 1000 * 1000;