+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# 1.7
+* Extend `GAIN_SEQ_BUFFER_SIZE_MAX` from 1024 to 2048
+* Add Gain mode option to `GainSequence`
+* Add `modulation::SineLegacy`
+* Update firmware version to 1.7
+
 # 1.6
 * Add `GainSequence`
 * Add `link::SOEM::set_lost_handler`